+++ lib/circuit_switch/builder.rb
@@ -46,7 +46,7 @@ module CircuitSwitch
         close_if_reach_limit: close_if_reach_limit,
         limit_count: limit_count,
         initially_closed: initially_closed,
-      }.select { |_, v| v }
+      }.reject { |_, v| v.nil? }
       assign_runner(**arguments)
       execute_run(&block)
     end
@@ -58,7 +58,7 @@ module CircuitSwitch
         stop_report_if: stop_report_if,
         stop_report_if_reach_limit: stop_report_if_reach_limit,
         limit_count: limit_count
-      }.select { |_, v| v }
+      }.reject { |_, v| v.nil? }
       assign_reporter(**arguments)
       execute_report
     end